refactor(lib): extract seededRandom into shared module
