+++ src/games/build-round-order.ts
@@ -4,21 +4,7 @@
  * When `seed` is provided, the shuffle is deterministic (seeded mulberry32).
  */
 
-const seededRandom = (seedStr: string): (() => number) => {
-  // djb2 hash → numeric seed
-  let h = 5381;
-  for (let i = 0; i < seedStr.length; i++) {
-    h = (Math.imul(h, 33) ^ (seedStr.codePointAt(i) ?? 0)) >>> 0;
-  }
-  // mulberry32 PRNG
-  let s = h;
-  return () => {
-    s = (s + 0x6d_2b_79_f5) >>> 0;
-    let t = Math.imul(s ^ (s >>> 15), 1 | s);
-    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t;
-    return ((t ^ (t >>> 14)) >>> 0) / 4_294_967_296;
-  };
-};
+import { seededRandom } from '@/lib/seeded-random';
 
 export function buildRoundOrder(
   length: number,
